ash-hf/requirements.txt
datasets
transformers
hf_xet
torch
tqdm
python-chess
modal
numpy
